Now the code generator is separated into smaller blocks, This is just a first step, once the code is mature enough it would be separated into interfaces that contain the logic and that may work for multiple generators.
